♻️ Refactor formatBytes
