+++ frontend/react/src/components/HardwareInfo/index.tsx
@@ -1,6 +1,7 @@
 import { useEffect, useState } from "react";
 import styles from "./styles.module.css";
 import { apiHandler } from "../../utils/apiHandler";
+import formatBytes from "../../utils/formatBytes";
 
 const HardwareInfoComponent: React.FC = () => {
   const [cpuInfo, setCpuInfo] = useState("0%");
@@ -23,27 +24,6 @@ const HardwareInfoComponent: React.FC = () => {
     setUsedStorage(data.usedStorage);
     setTotalStorage(data.totalStorage);
     setCpuClock(data.cpuClockSpeed);
-  }
-
-  function formatBytes(bytes: number, decimals = 2) {
-    if (!+bytes) return "0 Bytes";
-
-    const k = 1024;
-    const dm = decimals < 0 ? 0 : decimals;
-    const sizes = [
-      "Bytes",
-      "KiB",
-      "MiB",
-      "GiB",
-      "TiB",
-      "PiB",
-      "EiB",
-      "ZiB",
-      "YiB",
-    ];
-
-    const i = Math.floor(Math.log(bytes) / Math.log(k));
-    return `${parseFloat((bytes / Math.pow(k, i)).toFixed(dm))} ${sizes[i]}`;
   }
 
   function formatFrequency(mhz: number): string {
